I wasn't Proud of this
I wasn't Proud of how this came out quality wise. It has everything it needs for the assignment but I am not satisfied with how it looks so I updated it despite being after the date.

If you wish to see what it looked liked before the changes in the deadline use the main branch. But if you want to see a drastic improvement- and maybe mark that, use this